backend/requirements.txt
Django==3.2.16
django-colorfield==0.7.2
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.4.0
flake8==4.0.1
gunicorn==20.0.4
PyJWT==2.1.0
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2020.1
psycopg2-binary-2.9.3
sqlparse==0.3.1
typing_extensions==4.3.0
zipp==3.8.1